session check for subjects

+++ src/main/java/com/sk/skala/quizapi/service/SubjectService.java
@@ -10,6 +10,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import com.sk.skala.quizapi.config.Error;
+import com.sk.skala.quizapi.data.common.AccountInfo;
 import com.sk.skala.quizapi.data.common.PagedList;
 import com.sk.skala.quizapi.data.common.Response;
 import com.sk.skala.quizapi.data.table.Subject;
@@ -18,16 +19,20 @@ import com.sk.skala.quizapi.exception.ResponseException;
 import com.sk.skala.quizapi.repository.SubjectRepository;
 import com.sk.skala.quizapi.tools.StringTool;
 
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 
 @Service
 @RequiredArgsConstructor
-@Slf4j
 public class SubjectService {
 	private final SubjectRepository subjectRepository;
+	private final SessionHandler sessionHandler;
 
 	public Response getSubjectList(String name, int offset, int count) throws Exception {
+		if (!sessionHandler.isAdmin()) {
+			throw new ResponseException(Error.NOT_AUTHORIZED);
+		}
+
 		Pageable pageable = PageRequest.of(offset, count, Sort.by(Sort.Direction.ASC, "subjectName"));
 		Page<Subject> paged = subjectRepository.findAllBySubjectNameLike(StringTool.like(name), pageable);
 
@@ -43,8 +48,13 @@ public class SubjectService {
 		return response;
 	}
 
-	public Response getSubjectsByInstructor(Long id) throws Exception {
-		List<Subject> list = subjectRepository.findAllByInstructorId(id);
+	public Response getSubjectsByInstructor(Long instructorId) throws Exception {
+		AccountInfo account = sessionHandler.getAccountInfo();
+		if (account == null || account.getInstructorId() != instructorId) {
+			throw new ResponseException(Error.NOT_AUTHORIZED);
+		}
+
+		List<Subject> list = subjectRepository.findAllByInstructorId(instructorId);
 
 		PagedList pagedList = new PagedList();
 		pagedList.setTotal(list.size());
@@ -58,8 +68,9 @@ public class SubjectService {
 		return response;
 	}
 
+	@Transactional
 	public Response upsertSubject(Subject item) throws Exception {
-		log.info("{}", item.toString());
+
 		if (StringTool.isAnyEmpty(item.getSubjectName())) {
 			throw new ParameterException("subjectName");
 		}
@@ -68,21 +79,38 @@ public class SubjectService {
 			throw new ParameterException("instructorId");
 		}
 
+		AccountInfo account = sessionHandler.getAccountInfo();
+		if (!sessionHandler.isAdmin() || account == null || account.getInstructorId() != item.getInstructor().getId()) {
+			throw new ResponseException(Error.NOT_AUTHORIZED);
+		}
+
 		Optional<Subject> option = subjectRepository.findById(item.getId());
 		if (option.isEmpty()) {
 			item.setId(null);
 		}
-
 		subjectRepository.save(item);
+
+		List<Subject> list = subjectRepository.findAllByInstructorId(item.getInstructor().getId());
+		sessionHandler.storeAccessToken(account, list);
+
 		return new Response();
 	}
 
 	public Response deleteSubject(Subject item) {
+		AccountInfo account = sessionHandler.getAccountInfo();
+		if (!sessionHandler.isAdmin() || account == null || account.getInstructorId() != item.getInstructor().getId()) {
+			throw new ResponseException(Error.NOT_AUTHORIZED);
+		}
+
 		if (subjectRepository.existsById(item.getId())) {
 			subjectRepository.deleteById(item.getId());
 		} else {
 			throw new ResponseException(Error.DATA_NOT_FOUND);
 		}
+
+		List<Subject> list = subjectRepository.findAllByInstructorId(item.getInstructor().getId());
+		sessionHandler.storeAccessToken(account, list);
+
 		return new Response();
 	}
 }
